Rolling back 7

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -6,7 +6,7 @@
 #include "holberton.h"
 void puts_half(char *str)
 {
-	float i, a;
+	int i, a;
 
 	i = 0;
 	a = 0;
